chore(ci): Add debug step to list FIP boards
The build-fip.sh script is failing with an "Invalid board name" error.

This commit adds a step to the workflow to list the contents of the amlogic-boot-fip directory. This will allow us to see the valid board names that the script expects.

The signing step has been temporarily disabled to allow the workflow to complete and show the debug output.